worked on change requests

+++ scripts/test_al.sh
@@ -1,7 +1,13 @@
-export CUDA_VISIBLE_DEVICES=$1
+GPUNAME=$1
+if [[ -z $GPUNAME ]];
+then
+    echo `date`" - Missing mandatory arguments: GPU name. "
+    exit 1
+fi
+export CUDA_VISIBLE_DEVICES=$1GPUNAME
 rm -rf example_data/active/*
 rm .activecache.json
-python test_al.py \
+python scripts/run_pseudo_al.py \
 -di='example_data/images/' \
 -lf='example_data/labels.csv' \
 -df='example_data/_megadetector.json' \
